==> 02-LIWC.ipynb <==
This notebook prepares the data for Tableau exploratory data analysis from results of LIWC software.

import pandas as pd


selfimprovement = pd.read_csv("data/selfimpr_liwc.csv")
investing = pd.read_csv("data/investing_liwc.csv")
homeowners = pd.read_csv("data/homeowners_liwc.csv")


### Basic descriptive statistics

selfimprovement.moral.describe()

investing.moral.describe()

homeowners.moral.describe()

### Observe proportions of moral language across subreddits

Get total sizes

#Call shape method to get shape of dataframe and select the number of rows
total_si = selfimprovement.shape[0]
total_i = investing.shape[0]
total_h = homeowners.shape[0]

Get number of posts with moralization score higher than 0.25 (the LIWC's reported mean for their corpus of Reddit + other sources)

moral_si = selfimprovement[selfimprovement.moral > 0.25]
moral_i = investing[investing.moral > 0.25]
moral_h = homeowners[homeowners.moral > 0.25]

#Repear workflow to get number of rows
moral_si_n = moral_si.shape[0]
moral_i_n = moral_i.shape[0]
moral_h_n = moral_h.shape[0]

Calculate proportion of moralization language across the three subreddits

percentage_si = (moral_si_n / total_si) * 100
percentage_i = (moral_i_n / total_i) * 100
percentage_h = (moral_h_n / total_h) * 100

print(f"Proportion of moralized posts in the r/selfimprovement subreddit: {percentage_si:.2f}%")
print(f"Proportion of moralized posts in the r/investing subreddit:{percentage_i:.2f}%")
print(f"Proportion of moralized posts in the r/homeowners subreddit:{percentage_h:.2f}%")

## Feature engineering with foundations scores

Some feature engineering to create combined scores for each foundation, and combined scores for virtue and vice

def feature_engineering(df):

    # Calculate total for each foundation
    df["Care_total"] = df["Care_Virtue"] + df["Care_Vice"]
    df["Fairness_total"] = df["Fairness_Virtue"] + df["Fairness_Vice"]
    df["Loyalty_total"] = df["Loyalty_Virtue"] + df["Loyalty_Vice"]
    df["Authority_total"] = df["Authority_Virtue"] + df["Authority_Vice"]
    df["Sanctity_total"] = df["Sanctity_Virtue"] + df["Sanctity_Vice"]

    # Vice and virtue scores
    df["Virtue_total"] = (df["Care_Virtue"] + df["Fairness_Virtue"] 
                          + df["Loyalty_Virtue"] + df["Authority_Virtue"] 
                          + df["Sanctity_Virtue"])
    
    df["Vice_total"] = (df["Care_Vice"] + df["Fairness_Vice"] 
                    + df["Loyalty_Vice"] + df["Authority_Vice"] 
                    + df["Sanctity_Vice"])
    
    # Overall total score across all foundations
    df["Foundations_total_score"] = (
        df["Care_total"] + df["Fairness_total"] + df["Loyalty_total"] +
        df["Authority_total"] + df["Sanctity_total"])
    
    return df

selfimprovement2 = feature_engineering(selfimprovement)
investing2 = feature_engineering(investing)
homeowners2 = feature_engineering(homeowners)

Create one single df to use in Tableau

selfimprovement2["Subreddit"] = "selfimprovement"
investing2["Subreddit"] = "investing"
homeowners2["Subreddit"] = "homeowners"

all_reddits = pd.concat([selfimprovement2, investing2, homeowners2], ignore_index=True)

all_reddits

all_reddits.to_csv("data/engineered_morality.csv")

Ensure correct parsing

saved_csv = pd.read_csv("data/engineered_morality.csv")

print(all_reddits.shape)
print(saved_csv.shape)
