==> 04-pronouns-and-merging.ipynb <==
### Set up

import pandas as pd

file_df = pd.read_csv("data/engineered_morality.csv")
#Get only selfimprovement posts
data = file_df[file_df.Subreddit == "selfimprovement"]

Confirm correct size of data

data.shape

### Check distribution of pronouns variables

data.i.describe()

data.you.describe()

data.shehe.describe()

data.they.describe()

The distributions shows that there's more first person language than second or third language.

### Explore different cut offs points for categorizing posts by predominantly using first language vs not.

with pd.option_context('display.max_colwidth', None):
    display(data[data.i > 0.20].sort_values(by="i", ascending=True)[["id", "type", "cleaned_text","i", "shehe", "they"]][0:5])

At this point the posts are still not very "I" focused.

with pd.option_context('display.max_colwidth', None):
    display(data[data.i > 0.50].sort_values(by="i", ascending=True)[["id", "type", "cleaned_text","i", "shehe", "they"]][0:5])

Still not very "I" focused

with pd.option_context('display.max_colwidth', None):
    display(data[data.i > 2].sort_values(by="i", ascending=True)[["id", "type", "cleaned_text","i", "shehe", "they"]][0:5])

Still not very "I" focused

with pd.option_context('display.max_colwidth', None):
    display(data[data.i > 5.5].sort_values(by="i", ascending=True)[["id", "type", "cleaned_text","i", "shehe", "they"]][0:5])

5.5 looks like a good cut off point. Every post above it is gonna be categorized as predominanty first-person focused.

data["person_focus"] = data["i"].apply(lambda x: "First person" if x > 5.5 else "Not first person")

### See distribution

data.person_focus.value_counts()

Almost evenly distributed! Save in csv

data[["id", "person_focus"]].to_csv("data/selfimpr_by_person_focus.csv")

Ensure correct parsing

saved = pd.read_csv("data/selfimpr_by_person_focus.csv")
print(data.shape[0])
print(saved.shape[0])


Merge with topics (reduced) data and save as well

topics = pd.read_csv("data/selfimpr_topics.csv")[["id","dominant_topic"]]

merged = data.merge(topics, how="inner", on="id")

merged.to_csv("data/selfimpr_finaldata.csv")
