==> 05-plots.Rmd <==
---
title: "plots2"
author: "Natasha Carpio Castellanos"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

General set up

```{r}
library(ggplot2)
library(dplyr)
library(forcats)
library(readr)
library(ggtext)


# LOAD DATA
data <- read_csv("data/selfimpr_finaldata.csv")
```
Define topic labels and corresponding colors

```{r}
# DEFINE TOPIC LABELS
topic_mapping <- c(
  "1" = "Habits &\n Goals",
  "2" = "Education &\n Career",
  "3" = "Social\n life",
  "4" = "Health &\n Fitness",
  "5" = "Social\n media &\n Entertainment",
  "6" = "Confidence &\n Dating",
  "7" = "Family &\n Finances",
  "8" = "Learning",
  "9" = "Mental\n health"
)

# DEFINE COLORS
color_topics = c(
  "Habits &\n Goals" = "#E69F00",
  "Education &\n Career" = "#0072B2",
  "Social\n life" = "#56B4E9",
  "Health &\n Fitness" = "#009E73",
  "Social\n media &\n Entertainment" = "#F0E442",
  "Confidence &\n Dating" = "#D55E00",
  "Family &\n Finances" = "#CC79A7",
  "Learning" = "#999999",
  "Mental\n health" = "#660099"
)
```

Summarize morality scores

```{r}
average_morality <- data %>% 
  group_by(dominant_topic) %>%
  summarise(
    average_moral_liwc = mean(moral),
    average_mfd = mean(Foundations_total_score),
    average_Care_Virtue = mean(Care_Virtue),
    average_Care_Vice = mean(Care_Vice),
    average_Fairness_Virtue = mean(Fairness_Virtue),
    average_Fairness_Vice = mean(Fairness_Vice),
    average_Loyalty_Virtue = mean(Loyalty_Virtue),
    average_Loyalty_Vice = mean(Loyalty_Vice),
    average_Authority_Virtue = mean(Authority_Virtue),
    average_Authority_Vice = mean(Authority_Vice),
    average_Sanctity_Virtue = mean(Sanctity_Virtue),
    average_Sanctity_Vice = mean(Sanctity_Vice),
    average_Care_total = mean(Care_total),
    average_Fairness_total = mean(Fairness_total),
    average_Loyalty_total = mean(Loyalty_total),
    average_Authority_total = mean(Authority_total),
    average_Sanctity_total = mean(Sanctity_total),
    count = n()
  )
```

Get labels for topics in the summary table

```{r}
# Replace topic names
#average_morality <- average_morality %>%
 #mutate(topic_name = recode(dominant_topic, !!!topic_mapping)) %>%
 #reorder so topics with more morality appear at the top
 #mutate(topic_name = fct_reorder(topic_name, average_moral_liwc, .desc = FALSE))

average_morality <- average_morality %>%
 mutate(topic_name = recode((dominant_topic), !!!topic_mapping)) %>%
 mutate(topic_name = fct_reorder(topic_name, average_moral_liwc, .desc = FALSE))
```

```{r}
reference_mean = 0.25

fig1 <- ggplot(average_morality, aes(x = average_moral_liwc, y = factor(topic_name), size = count, color = as.character(topic_name))) +
  geom_point(alpha = 0.6, position = position_jitter(height = 0.2, width = 0)) +  
  geom_vline(xintercept=reference_mean, color="red", linetype ="dashed",size=0.5) +
  scale_size_continuous(range = c(10, 25)) +
  scale_color_manual(values = color_topics) +  
  labs(title = "Moral Expression Across Topics",
       subtitle = "(Size proportional to percentage of corpus)",
       x = "Avg. Moral LIWC Score compared to <span style='color:red;'>- -LIWC testing corpus baseline- -</span>",
       y = "Topics") +
  theme_bw() +
    theme(
    plot.title = element_text(size = 20, hjust=0.5),  
    plot.subtitle = element_text(size = 10, hjust=0.5),
    axis.title.x = element_markdown(size = 10),            
    axis.title.y = element_text(size = 10),
    axis.text.x = element_text(size = 9),              
    axis.text.y = element_text(size = 9),
    legend.position = "none")

fig1
```

Same moralization across topics but filtering for only posts in first person

```{r}
firstperson_data <- data %>% filter(person_focus == "First person")

average_morality <- firstperson_data %>% 
  group_by(dominant_topic) %>%
  summarise(
    average_moral_liwc = mean(moral),
    average_mfd = mean(Foundations_total_score),
    average_Care_Virtue = mean(Care_Virtue),
    average_Care_Vice = mean(Care_Vice),
    average_Fairness_Virtue = mean(Fairness_Virtue),
    average_Fairness_Vice = mean(Fairness_Vice),
    average_Loyalty_Virtue = mean(Loyalty_Virtue),
    average_Loyalty_Vice = mean(Loyalty_Vice),
    average_Authority_Virtue = mean(Authority_Virtue),
    average_Authority_Vice = mean(Authority_Vice),
    average_Sanctity_Virtue = mean(Sanctity_Virtue),
    average_Sanctity_Vice = mean(Sanctity_Vice),
    average_Care_total = mean(Care_total),
    average_Fairness_total = mean(Fairness_total),
    average_Loyalty_total = mean(Loyalty_total),
    average_Authority_total = mean(Authority_total),
    average_Sanctity_total = mean(Sanctity_total),
    count = n()
  )

# Replace topic names
average_morality <- average_morality %>%
  mutate(topic_name = recode(dominant_topic, !!!topic_mapping)) %>%
  # reorder so topics with more morality appear at the top
  mutate(topic_name = fct_reorder(topic_name, average_moral_liwc, .desc = FALSE))

reference_mean = 0.25

fig2 <- ggplot(average_morality, aes(x = average_moral_liwc, y = factor(topic_name), size = count, color = topic_name)) +
  geom_point(alpha = 0.6, position = position_jitter(height = 0.2, width = 0)) +  
  geom_vline(xintercept=reference_mean, color="red", linetype ="dashed",size=0.5) +
  scale_size_continuous(range = c(10, 25)) +
  scale_color_manual(values = color_topics) +  
  labs(title = "Moral Expression Across Topics",
       subtitle = "(Size proportional to percentage of corpus)",
       x = "Avg. Moral LIWC Score compared to <span style='color:red;'>- -LIWC testing corpus baseline- -</span>",
       y = "Topics") +
  theme_bw() +
    theme(
    plot.title = element_text(size = 20, hjust=0.5),  
    plot.subtitle = element_text(size = 10, hjust=0.5),
    axis.title.x = element_markdown(size = 10),            
    axis.title.y = element_text(size = 10),
    axis.text.x = element_text(size = 9),              
    axis.text.y = element_text(size = 9),
    legend.position = "none")

fig2
```
```{r}
ggsave("plots/moralization-topics.png", plot = fig1, width = 8, height = 6, dpi = 300)
ggsave("plots/firstperson-moralization-topics.png", plot = fig2, width = 8, height = 6, dpi = 300)
```


Moral Foundations Theory plots

```{r}
# BASE PLOT FUNCTION
make_foundation_plot <- function(data, y_var, title) {
  ggplot(data, aes(x = topic_name, y = .data[[y_var]], fill = topic_name)) +
    geom_bar(stat = "identity", show.legend = FALSE) +
    scale_fill_manual(values = color_topics) + 
    labs(title = title, x = "Topics", y = "Score") +
    theme_minimal() +
    theme(
      plot.title = element_text(size = 20, hjust = 0.5),
      axis.title.x = element_text(size = 10),
      axis.title.y = element_text(size = 10),
      axis.text.x = element_text(size = 9),
      axis.text.y = element_text(size = 9),
      plot.margin = margin(t = 10, r = 10, b = 10, l = 60)
    ) +
    scale_y_continuous(limits = c(0, 1.6)) +
    coord_flip() -> plot
}

# SHARED DATA
plot_data_summary <- data %>%
  select(dominant_topic, Sanctity_total, Care_total) %>%
  group_by(dominant_topic) %>%
  summarise(Sanctity_total = mean(Sanctity_total, na.rm = TRUE),
            Care_total = mean(Care_total, na.rm = TRUE)) %>%
  mutate(topic_name = recode(dominant_topic, !!!topic_mapping))

# SANCTITY: reorder by Sanctity
plot_data_sanctity <- plot_data_summary %>%
  mutate(topic_name = fct_reorder(topic_name, Sanctity_total, .desc = FALSE))

# CARE: reorder by Care
plot_data_care <- plot_data_summary %>%
  mutate(topic_name = fct_reorder(topic_name, Care_total, .desc = FALSE))

# GENERATE PLOTS
sanctity <- make_foundation_plot(plot_data_sanctity, "Sanctity_total", "Total Sanctity language")
care <- make_foundation_plot(plot_data_care, "Care_total", "Total Care language")

```

```{r}
care
sanctity
```

```{r}
ggsave("plots/sanctity_total_plot.png", plot = sanctity, width = 8, height = 6, dpi = 300)
ggsave("plots/care_total_plot.png", plot = care, width = 8, height = 6, dpi = 300)
```
