==> 08-moralBERT-processing.ipynb <==
import pandas as pd

selfimprovement = pd.read_csv("data/moralBERT-results/MoralBERTresults400-selfimprovement.csv")
investing = pd.read_csv("data/moralBERT-results/MoralBERTresults400-investing.csv")
homeowners = pd.read_csv("data/moralBERT-results/MoralBERTresults400-homeowners.csv")

investing

selfimprovement

homeowners

selfimprovement["subreddit"] = "selfimprovement"
investing["subreddit"] = "investing"
homeowners["subreddit"] = "homeowners"

df = pd.concat([selfimprovement, investing, homeowners], ignore_index=True)
df

foundations = ["care", "harm", "fairness", "cheating", "loyalty", "betrayal", "authority", "subversion", "purity", "degradation"]

print(df.dtypes)

def threshold_function(value):
    if value > 0.3:
        return 1
    else:
        return 0

# Apply the threshold function to each cell in the DataFrame
for column in df.columns[1:-1]:  # Exclude the 'sentence' column
    df[column] = df[column].apply(threshold_function)


df

Summarize into each foundation

df['care_dim'] = ((df['care'] == 1) | (df['harm'] == 1)).astype(int)
df['fairness_dim'] = ((df['fairness'] == 1) | (df['cheating'] == 1)).astype(int)
df['loyalty_dim'] = ((df['loyalty'] == 1) | (df['betrayal'] == 1)).astype(int)
df['authority_dim'] = ((df['authority'] == 1) | (df['subversion'] == 1)).astype(int)
df['sanctity_dim'] = ((df['purity'] == 1) | (df['degradation'] == 1)).astype(int)



df.to_csv("data/moralBERT-results/moralBERT-binary400.csv")
